+++ RELEASENOTES.md
@@ -59,3 +59,11 @@
 * CB-6465: Add license headers to Tizen code
 * CB-6460: Update license headers
 * Add NOTICE file
+
+### 0.3.9 (Jun 05, 2014)
+* updated notice file
+* Github close #11
+* Extended vibrateWithPattern to allow for pattern repetition, implemented a complementary cancelVibration function and adapted documentation.
+* Implemented vibrateWithPattern (for android) and adapted documentation.
+* CB-6811 Add license to CONTRIBUTING.md
+* CB-6491 add CONTRIBUTING.md